Abstract validator status logger to an interface
