+++ validator/client/src/main/java/tech/pegasys/teku/validator/client/ValidatorStatusLogger.java
@@ -13,122 +13,18 @@
 
 package tech.pegasys.teku.validator.client;
 
-import static com.google.common.base.Preconditions.checkArgument;
-import static tech.pegasys.teku.infrastructure.logging.StatusLogger.STATUS_LOG;
+public interface ValidatorStatusLogger {
 
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Optional;
-import java.util.Set;
-import java.util.concurrent.atomic.AtomicInteger;
-import java.util.concurrent.atomic.AtomicReference;
-import tech.pegasys.teku.api.response.v1.beacon.ValidatorStatus;
-import tech.pegasys.teku.bls.BLSPublicKey;
-import tech.pegasys.teku.validator.api.ValidatorApiChannel;
+  ValidatorStatusLogger NOOP =
+      new ValidatorStatusLogger() {
+        @Override
+        public void printInitialValidatorStatuses() {}
 
-public class ValidatorStatusLogger {
+        @Override
+        public void checkValidatorStatusChanges() {}
+      };
 
-  private static final int VALIDATOR_KEYS_PRINT_LIMIT = 20;
+  void printInitialValidatorStatuses();
 
-  final Set<BLSPublicKey> validatorPublicKeys;
-  final ValidatorApiChannel validatorApiChannel;
-  final AtomicReference<Map<BLSPublicKey, ValidatorStatus>> latestValidatorStatuses =
-      new AtomicReference<>();
-
-  public ValidatorStatusLogger(
-      Set<BLSPublicKey> validatorPublicKeys, ValidatorApiChannel validatorApiChannel) {
-    checkArgument(!validatorPublicKeys.isEmpty());
-    this.validatorPublicKeys = validatorPublicKeys;
-    this.validatorApiChannel = validatorApiChannel;
-  }
-
-  public void printInitialValidatorStatuses() {
-    validatorApiChannel
-        .getValidatorStatuses(validatorPublicKeys)
-        .thenAccept(
-            maybeValidatorStatuses -> {
-              if (maybeValidatorStatuses.isEmpty()) {
-                STATUS_LOG.unableToRetrieveValidatorStatusesFromBeaconNode();
-                return;
-              }
-
-              Map<BLSPublicKey, ValidatorStatus> validatorStatuses = maybeValidatorStatuses.get();
-              if (validatorPublicKeys.size() < VALIDATOR_KEYS_PRINT_LIMIT) {
-                printValidatorStatusesOneByOne(validatorStatuses);
-              } else {
-                printValidatorStatusSummary(validatorStatuses);
-              }
-            })
-        .reportExceptions();
-  }
-
-  private void printValidatorStatusesOneByOne(
-      Map<BLSPublicKey, ValidatorStatus> validatorStatuses) {
-    for (BLSPublicKey publicKey : validatorPublicKeys) {
-      Optional<ValidatorStatus> maybeValidatorStatus =
-          Optional.ofNullable(validatorStatuses.get(publicKey));
-      maybeValidatorStatus.ifPresentOrElse(
-          validatorStatus ->
-              STATUS_LOG.validatorStatus(validatorStatus.name(), publicKey.toAbbreviatedString()),
-          () -> STATUS_LOG.unableToRetrieveValidatorStatus(publicKey.toAbbreviatedString()));
-    }
-  }
-
-  private void printValidatorStatusSummary(Map<BLSPublicKey, ValidatorStatus> validatorStatuses) {
-    Map<ValidatorStatus, AtomicInteger> validatorStatusCount = new HashMap<>();
-    final AtomicInteger unknownValidatorCountReference = new AtomicInteger(0);
-    for (BLSPublicKey publicKey : validatorPublicKeys) {
-      Optional<ValidatorStatus> maybeValidatorStatus =
-          Optional.ofNullable(validatorStatuses.get(publicKey));
-      maybeValidatorStatus.ifPresentOrElse(
-          status -> {
-            AtomicInteger count =
-                validatorStatusCount.computeIfAbsent(status, __ -> new AtomicInteger(0));
-            count.incrementAndGet();
-          },
-          unknownValidatorCountReference::incrementAndGet);
-    }
-
-    for (Map.Entry<ValidatorStatus, AtomicInteger> statusCount : validatorStatusCount.entrySet()) {
-      STATUS_LOG.validatorStatusSummary(statusCount.getValue().get(), statusCount.getKey().name());
-    }
-
-    final int unknownValidatorCount = unknownValidatorCountReference.get();
-    if (unknownValidatorCount > 0) {
-      STATUS_LOG.unableToRetrieveValidatorStatusSummary(unknownValidatorCountReference.get());
-    }
-  }
-
-  public void checkValidatorStatusChanges() {
-    validatorApiChannel
-        .getValidatorStatuses(validatorPublicKeys)
-        .thenAccept(
-            maybeNewValidatorStatuses -> {
-              if (maybeNewValidatorStatuses.isEmpty()) {
-                STATUS_LOG.unableToRetrieveValidatorStatusesFromBeaconNode();
-                return;
-              }
-
-              Map<BLSPublicKey, ValidatorStatus> newValidatorStatuses =
-                  maybeNewValidatorStatuses.get();
-
-              Map<BLSPublicKey, ValidatorStatus> oldValidatorStatuses =
-                  latestValidatorStatuses.getAndSet(newValidatorStatuses);
-              if (oldValidatorStatuses == null) {
-                return;
-              }
-
-              for (BLSPublicKey key : oldValidatorStatuses.keySet()) {
-                ValidatorStatus oldStatus = oldValidatorStatuses.get(key);
-                ValidatorStatus newStatus = newValidatorStatuses.get(key);
-                if (oldStatus.equals(newStatus)) {
-                  continue;
-                }
-
-                STATUS_LOG.validatorStatusChange(
-                    oldStatus.name(), newStatus.name(), key.toAbbreviatedString());
-              }
-            })
-        .reportExceptions();
-  }
+  void checkValidatorStatusChanges();
 }
